move environments + wrappers to one module
It didn't make sense to keep the environments in rollouts to begin with. When I wanted a wrapper to depend on an environment base class, it had to import the rollouts module (which makes no sense). Now, environments are factored out much more naturally. The one disadvantage of this is that the ThreadedRoller depends on an environment API (so not completely decoupled), so I deleted ThreadedRoller for good measure (it was never really good for anything, anyway).