fix beta log probs
Essentially, numpy sampling seems to give us values on the border sometimes (e.g. 0 or 1). I'm not sure why this is--it's probably due to some crude approximation--but that's a story for another day.

Anyway, tf gives a -inf logprob at 0 or 1. We can get around this by forcing the values to be slightly pushed in at the margins (by epsilon=1e-20, in this case).