core DQN code
This doesn't yet include a training loop, but it does compute the loss terms and feed dicts for those terms.

Some things to nail down before doing a training loop:
 - how to deal with epsilon-greedy exploration/schedules
 - how to periodically evaluate the model
 - how to deal with replay buffers & priority schedules

The above makes it seem like the training loop will take a LOT of arguments. It would be nice to get things split up into several different methods, but I'm not sure of the best way to do that.